requirements.txt
discord_webhook==1.4.1
Django==5.0.9
google_api_python_client==2.137.0
google_auth_oauthlib==1.2.1
Pillow==12.0.0
protobuf==6.33.0
python-dotenv==1.2.1
Requests==2.32.5
django-browser-reload==1.21.0
djangorestframework==3.15.2
drf-yasg==1.21.8
django-cors-headers==4.9.0
supabase
djangorestframework-simplejwt